requirements.txt
pyyaml
jinja2
docker
runpod
huggingface-hub
